Added tools to CG operator dedicated to CB reconstruction for performing a support constraint. The constraint is set as an attribute image which is a binary mask (1 in the support and 0 outside). Is the mask is set, the constraint operator consists in applying the mask P in the CG normal equations (AX=B) which become :
P^t A P X = P^t B

Applying the mask is implemented a protected method of the filter and is injected in the pipeline. To avoid many conditional cases implementation ("if preconditioned", "if regularized"), the method takes an image in argument and perform the calculation only if the mask is given. If not the method return the input image.