Run test on appveyor. AppVeyor config cleaning
Add support to run tests on AppVeyor.
Add dependency on appveyor.yml for ITK caching. This should prevent update
of ITK cache if something else than this file is modified.
Because RTK writes some outputs into ITK binary directory, we don't want
the cache to be uploaded. If it tries to upload the cache, it fails because
the build size is too large. This results in caching failing and RTK not
being built on the next triggered build (because next build will build ITK
and update the cache).