Wrap new classes for Python
Adding new .wrap files to define templates used for python wrapping.

A few things have to be considered for future commits :

- The use of ConstPointer type will generate warnings when wrapping for
python. This is because itk::Image and itk::VectorImage are not wrapped
using the wrap method POINTER_WITH_CONST_POINTER. A solution is either to
modify the wrap method used for those classes in ITK, or to use "const *"
instead of ConstPointer in RTK. For now, the warnings fired because of const
pointers should not alter the wrapping process.

- Add requirement for ITK_WRAP_unsigned_short to be ON in both ITK and RTK.
We might also require image dimension 1 to be wrapped. Classes requiring
1D images are relatively rare for now so we'll leave it like this and prevent
those classes from being wrapped. (See Reg1DExtractShroudSignalImageFilter.notwrap
and DPExtractShroudSignalImageFilter.notwrap).

- GeneralPurposeFunction should all be static to avoid redefinition.
For now, only the required one as been turned to static.