BUG: Force wrapping of ExtractPhaseImageFilter
When using itk_wrap_image_filter with dimension restriction, the
macro verifies that the dimension parameter is included in CMake
ITK_WRAP_IMAGE_DIMS variable.
Because RTK handle wrapping of dimension 1 regardless of
ITK_WRAP_IMAGE_DIMS value, this filter should not use the macro if we
want to ensure that it is wrapped for Python.