ENH: Compute the backprojection of ones only once in OSEM
A backprojection of a volume filled with ones is needed for the OSEM algorithm.
Until now, this backprojection was computed at every iteration. However,
the result is independent of the iterations. Thus, computing this
backprojection only during the first iteration and reused the result for
the next is enough and should reduce the computation time of the OSEM
algorithm.