ENH: Add superior and inferior clip images for Joseph-based forward projector
The superior and inferior clip images define for each pixel of the
projections the value of the superior and inferior clip of the ray
emitted from that pixel. Each ray is clipped from
source+inferior_clip*(pixel-source) to
source+superior_clip*(pixel-source) with inferior_clip and superior_clip
equal 0 and 1 by default.