ENH: Add conversion from the attenuated to exponential Radon transform
The conversion is explained p47 of Natterer's book, "The mathematics of
computerized tomography", 1986. The conversion assumes that the emission
is contained in a K region (referred to as $\Omega$ in the book) of
constant attenuation $\mu_0$. The projector therefore uses a mask as
third input with the same voxel lattice as the attenuation map in the
first input. $\mu_0$ is calculated as the average in the K region.