BUG: Fix FDK variance reconstruction
The output of the rtkFDKVarianceReconstructionFilter was full of nan
values. The bug was introduced by the last force-pushed of the pull
request #632 and not detected by the test because too few projections
were used. This commit fix the bug and add more projections to the test
in order to potentially detect futur bug.