BUG: Avoid numerical error in quadric ray intersection computation
The error encountered in #819 is due to numerical errors when testing
whether the segment defined by the two intersections of the ray with the
quadric is in or out of the quadric. A new strategy, computationally
simpler and hopefully less prone to numerical errors, is used: test
whether the ray origin is between these points and in the quadric.